@@ -11,67 +11,68 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.7",
-    "@agm/snazzy-info-window": "^1.0.0-beta.7",
-    "@angular/animations": "^8.1.3",
-    "@angular/cdk": "^8.1.1",
-    "@angular/common": "^8.1.3",
-    "@angular/compiler": "^8.1.3",
-    "@angular/core": "^8.1.3",
+    "@agm/core": "^1.0.0",
+    "@agm/snazzy-info-window": "^1.0.0",
+    "@angular/animations": "^8.2.11",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.11",
+    "@angular/compiler": "^8.2.11",
+    "@angular/core": "^8.2.11",
     "@angular/fire": "^5.2.1",
-    "@angular/flex-layout": "^8.0.0-beta.26",
-    "@angular/forms": "^8.1.3",
-    "@angular/material": "^8.1.1",
-    "@angular/platform-browser": "^8.1.3",
-    "@angular/platform-browser-dynamic": "^8.1.3",
-    "@angular/router": "^8.1.3",
+    "@angular/flex-layout": "^8.0.0-beta.27",
+    "@angular/forms": "^8.2.11",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.11",
+    "@angular/platform-browser-dynamic": "^8.2.11",
+    "@angular/router": "^8.2.11",
     "angular2-text-mask": "^9.0.0",
-    "echarts": "^4.2.1",
+    "echarts": "^4.4.0",
     "file-saver": "^2.0.2",
-    "firebase": "^6.3.3",
-    "firebase-tools": "^7.2.1",
+    "firebase": "^7.2.1",
+    "firebase-tools": "^7.6.1",
     "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "howler": "^2.1.2",
+    "intro.js": "^2.9.3",
     "jquery": "^3.4.1",
     "jquery-slimscroll": "^1.3.8",
     "moment": "^2.24.0",
     "ng5-slider": "^1.2.4",
     "ngx-echarts": "^4.2.1",
     "ngx-strongly-typed-forms": "^8.0.1",
-    "rxjs": "^6.5.2",
-    "saturn-datepicker": "^7.4.0",
-    "stacktrace-js": "^2.0.0",
+    "rxjs": "^6.5.3",
+    "saturn-datepicker": "^8.0.1",
+    "stacktrace-js": "^2.0.1",
     "tslib": "^1.9.0",
-    "xlsx": "^0.14.4",
-    "zone.js": "^0.10.0"
+    "xlsx": "^0.15.1",
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.801.2",
-    "@angular/cli": "^8.1.2",
-    "@angular/compiler-cli": "^8.1.3",
-    "@angular/language-service": "^8.1.3",
+    "@angular-devkit/build-angular": "^0.803.13",
+    "@angular/cli": "^8.3.13",
+    "@angular/compiler-cli": "^8.2.11",
+    "@angular/language-service": "^8.2.11",
     "@angularclass/hmr": "^2.1.3",
     "@angularclass/hmr-loader": "^3.0.4",
     "@types/core-js": "^2.5.2",
-    "@types/echarts": "^4.1.9",
-    "@types/googlemaps": "^3.37.0",
+    "@types/echarts": "^4.4.0",
+    "@types/googlemaps": "^3.38.0",
     "@types/hammerjs": "^2.0.34",
-    "@types/jasmine": "^3.3.16",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "^12.6.8",
+    "@types/jasmine": "^3.4.4",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^12.11.5",
     "angular-router-loader": "^0.8.5",
     "angular2-template-loader": "^0.6.2",
-    "autoprefixer": "^9.6.1",
+    "autoprefixer": "^9.6.5",
     "awesome-typescript-loader": "^5.0.0",
     "bootstrap": "^4.1.1",
-    "codelyzer": "^5.1.0",
-    "css-loader": "^3.1.0",
-    "jasmine-core": "^3.4.0",
+    "codelyzer": "^5.2.0",
+    "css-loader": "^3.2.0",
+    "jasmine-core": "^3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "json-loader": "^0.5.4",
-    "karma": "^4.2.0",
-    "karma-chrome-launcher": "^3.0.0",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
     "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-jasmine": "^2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
