@@ -1,12 +1,12 @@
-import {Injectable} from "@angular/core";
-import {AngularFireMessaging} from "@angular/fire/messaging";
-import {BehaviorSubject} from "rxjs";
-import {AngularFirestore} from "@angular/fire/firestore";
-import {Admin_} from "../../models/Admin";
-import {NotificationService} from "../../shared/services/notification.service";
-import {FCM} from "../../models/FCM";
-import {distinctUntilChanged} from "rxjs/operators";
-import {AdminsService} from "./admins.service";
+import { Injectable } from "@angular/core";
+import { AngularFireMessaging } from "@angular/fire/messaging";
+import { BehaviorSubject } from "rxjs";
+import { AngularFirestore } from "@angular/fire/firestore";
+import { Admin } from "../../models/Admin";
+import { NotificationService } from "../../shared/services/notification.service";
+import { FCM } from "../../models/FCM";
+import { distinctUntilChanged } from "rxjs/operators";
+import { AdminsService } from "./admins.service";
 
 @Injectable({
   providedIn: "root"
@@ -15,9 +15,9 @@ export class FcmService {
   currentMessage = new BehaviorSubject(null);
 
   constructor(private db: AngularFirestore,
-              private afMessaging: AngularFireMessaging,
-              private notification: NotificationService,
-              private adminservice: AdminsService) {
+    private afMessaging: AngularFireMessaging,
+    private notification: NotificationService,
+    private adminservice: AdminsService) {
     this.adminservice.observableuserdata
       .pipe(distinctUntilChanged())
       .subscribe(admin => {
@@ -28,13 +28,13 @@ export class FcmService {
       });
   }
 
-  updateusertokes(user: Admin_, token) {
+  updateusertokes(user: Admin, token) {
     if (user.fcmtokens.web !== token) {
-      this.db.firestore.collection("admins").doc(user.Id).update({"fcmtokens.web": token});
+      this.db.firestore.collection("admins").doc(user.Id).update({ "fcmtokens.web": token });
     }
   }
 
-  requestPermission(user: Admin_) {
+  requestPermission(user: Admin) {
     this.afMessaging.requestToken.subscribe(
       (token) => {
         this.updateusertokes(user, token);
@@ -50,21 +50,21 @@ export class FcmService {
       (payload: FCM) => {
         // console.log('new message received. ', payload);
         switch (payload.notification.title) {
-          case "Payment Received" :
+          case "Payment Received":
             return this.notification.notify({
               alert_type: "cash",
               duration: 5000,
               title: payload.notification.title,
               body: payload.notification.body
             });
-          case "Unprocessed Payment" :
+          case "Unprocessed Payment":
             return this.notification.notify({
               alert_type: "unprocessedpayment",
               duration: 5000,
               title: payload.notification.title,
               body: payload.notification.body
             });
-          default :
+          default:
             return null;
         }
       });
