@@ -1,12 +1,12 @@
 // import { BatchSelectorComponent } from './../batch-selector/batch-selector.component';
-import {Component, Input, OnDestroy, OnInit} from "@angular/core";
-import {MatDialog} from "@angular/material"; //added dialog data receive
-import {NotificationService} from "../../shared/services/notification.service";
-import {AngularFirestore} from "@angular/fire/firestore";
-import {Truck_} from "../../models/Truck";
-import {Order_} from "../../models/Order";
-import {TrucksService} from "../services/trucks.service";
-import {ReplaySubject} from "rxjs";
+import { Component, Input, OnDestroy, OnInit } from "@angular/core";
+import { MatDialog } from "@angular/material"; // added dialog data receive
+import { NotificationService } from "../../shared/services/notification.service";
+import { AngularFirestore } from "@angular/fire/firestore";
+import { Truck } from "../../models/Truck";
+import { Order } from "../../models/Order";
+import { TrucksService } from "../services/trucks.service";
+import { ReplaySubject } from "rxjs";
 
 @Component({
   selector: "order-details",
@@ -19,9 +19,9 @@ export class OrderDetailsComponent implements OnInit, OnDestroy {
   position1 = "left";
   position2 = "above";
 
-  truck: Truck_;
+  truck: Truck;
   displayedColumns = ["Id", "Company", "Contact", "Time", "Phone", "PMS", "AGO", "IK", "Total", "Action", "Status"];
-  @Input() order: Order_;
+  @Input() order: Order;
 
   /**
    * this keeps a local copy of all the subscriptions within this service
@@ -30,14 +30,14 @@ export class OrderDetailsComponent implements OnInit, OnDestroy {
   comopnentDestroyed: ReplaySubject<boolean> = new ReplaySubject<boolean>();
 
   constructor(private dialog: MatDialog,
-              private db: AngularFirestore,
-              private truckservice: TrucksService,
-              private notificationService: NotificationService) {
+    private db: AngularFirestore,
+    private truckservice: TrucksService,
+    private notificationService: NotificationService) {
     if (!this.order) {
       return;
     }
-    let subscription = this.truckservice.getTruck(this.order.Id).onSnapshot(truckdata => {
-      this.truck = truckdata.data() as Truck_;
+    const subscription = this.truckservice.getTruck(this.order.Id).onSnapshot(truckdata => {
+      this.truck = truckdata.data() as Truck;
     });
     this.subscriptions.set(`statsrange`, subscription);
 
