@@ -1,15 +1,15 @@
-import {Injectable} from '@angular/core';
-import {BehaviorSubject} from "rxjs";
-import {emptyorder, Order_} from "../../models/Order";
+import { Injectable } from "@angular/core";
+import { BehaviorSubject } from "rxjs";
+import { emptyorder, Order } from "../../models/Order";
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: "root"
 })
 /**
  * this holds variables that are set in one component but are necessary for usage in another component
  */
 export class ComponentCommunicationService {
-  clickedorder: BehaviorSubject<Order_> = new BehaviorSubject<Order_>({...emptyorder});
+  clickedorder: BehaviorSubject<Order> = new BehaviorSubject<Order>({ ...emptyorder });
   truckDeleted: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
 
   constructor() {
