@@ -1,6 +1,6 @@
 import { Component, EventEmitter, OnDestroy, OnInit } from "@angular/core";
-import { emptytruck, Truck_ } from "../../models/Truck";
-import { Order_ } from "../../models/Order";
+import { emptytruck, Truck } from "../../models/Truck";
+import { Order } from "../../models/Order";
 import { TrucksService } from "../services/trucks.service";
 import * as moment from "moment";
 import { ConfirmDialogComponent } from "../confirm-dialog/confirm-dialog.component";
@@ -23,8 +23,8 @@ import { firestore } from "firebase";
   styleUrls: ["./truck-details.component.scss"]
 })
 export class TruckDetailsComponent implements OnInit, OnDestroy {
-  order: Order_;
-  clickedtruck: Truck_ = { ...emptytruck };
+  order: Order;
+  clickedtruck: Truck = { ...emptytruck };
   loadingtruck = false;
   position = "above";
 
@@ -135,7 +135,7 @@ export class TruckDetailsComponent implements OnInit, OnDestroy {
     }
     const subscription = this.truckservice.getTruck(this.order.Id)
       .onSnapshot(truck => {
-        this.clickedtruck = truck.data() as Truck_;
+        this.clickedtruck = truck.data() as Truck;
         if (truck.exists) {
           this.clickedtruck.Id = truck.id;
         } else {
@@ -172,7 +172,7 @@ export class TruckDetailsComponent implements OnInit, OnDestroy {
    *
    * @param truck
    */
-  freezetruck(truck: Truck_) {
+  freezetruck(truck: Truck) {
     const dialogRef = this.dialog.open(ConfirmDialogComponent,
       {
         role: "dialog",
@@ -278,7 +278,7 @@ export class TruckDetailsComponent implements OnInit, OnDestroy {
    *
    * @param truck
    */
-  resetTruck(truck: Truck_) {
+  resetTruck(truck: Truck) {
     const dialogRef = this.dialog.open(ConfirmDialogComponent,
       {
         role: "dialog",
