@@ -1,6 +1,6 @@
-import {async, TestBed} from "@angular/core/testing";
-import {RouterTestingModule} from "@angular/router/testing";
-import {AppComponent} from "./app.component";
+import { async, TestBed } from "@angular/core/testing";
+import { RouterTestingModule } from "@angular/router/testing";
+import { AppComponent } from "./app.component";
 
 describe("AppComponent", () => {
   beforeEach(async(() => {
@@ -20,16 +20,16 @@ describe("AppComponent", () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'emkay-beta'`, () => {
+  it(`should have as title 'daudi'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.debugElement.componentInstance;
-    expect(app.title).toEqual("emkay-beta");
+    expect(app.title).toEqual("daudi");
   });
 
   it("should render title in a h1 tag", () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector("h1").textContent).toContain("Welcome to emkay-beta!");
+    expect(compiled.querySelector("h1").textContent).toContain("Welcome to daudi!");
   });
 });
