@@ -1,46 +1,51 @@
 import { inituser, User } from "./universal";
 
-export interface Admin_ {
-  Active: boolean,
+export interface Admin {
+  Active: boolean;
   qbconfig: {
     companyid: string,
     QbId: string,
     sandbox: boolean
-  }
-  Id: string,
+  };
+  Id: string;
   config: {
     level: string,
     viewsandbox: boolean,
-    depotid: string,
+    app: {
+      depotid: string,
+    },
+    fcm: {
+      tokens: {
+        web: string,
+        apk: string
+      };
+      settings: {
+        truck: {
+          live: boolean,
+          sandbox: boolean
+        },
+        payment: {
+          live: boolean,
+          sandbox: boolean
+        }
+      },
+    };
     approvedby: User
     /**
      * This keeps a reference to the id of the user type
      */
     type: number,
-  },
-  data: {
+  };
+  status: {
+    online: boolean,
+    time: Date
+  };
+
+  profile: {
     email: string,
     uid: string,
     photoURL: string,
     name: string,
-  },
-  status: {
-    online: boolean,
-    time: Date
-  },
-  settings: {
-    fcm: {
-      trueck: {
-        live: boolean,
-        sandbox: boolean
-      },
-      payment: {
-        live: boolean,
-        sandbox: boolean
-      },
-    },
-  }
-  profiledata: {
     address: {
       Id: string,
       Line1: string,
@@ -52,14 +57,11 @@ export interface Admin_ {
     dob: string,
     bio: string,
     phone: string,
-  },
-  fcmtokens: {
-    web: string,
-    apk: string
   };
+
 }
 
-export const emptyadmin: Admin_ = {
+export const emptyadmin: Admin = {
   Active: null,
   qbconfig: {
     companyid: null,
@@ -70,43 +72,37 @@ export const emptyadmin: Admin_ = {
   config: {
     viewsandbox: false,
     level: null,
-    approvedby: inituser,
-    depotid: null,
+    approvedby: { ...inituser },
+    app: {
+      depotid: null
+    },
+    fcm: {
+      settings: {
+        payment: null,
+        truck: null
+      },
+      tokens: {
+        apk: null,
+        web: null
+      }
+    },
     type: null,
   },
-  data: {
-    email: null,
-    uid: null,
-    photoURL: null,
-    name: null
-  },
+
   status: {
     online: false,
     time: null
   },
-  profiledata: {
-    gender: null,
+  profile: {
     address: null,
-    dob: null,
     bio: null,
-    phone: null
-  },
-  fcmtokens: {
-    web: null,
-    apk: null
-  },
-  settings: {
-    fcm: {
-      trueck: {
-        live: false,
-        sandbox: false
-      },
-      payment: {
-        live: false,
-        sandbox: false
-      }
-
-    }
+    dob: null,
+    email: null,
+    gender: null,
+    name: null,
+    phone: null,
+    photoURL: null,
+    uid: null
   }
 
 };
