@@ -1,11 +1,11 @@
 export interface FCM {
-  collapse_key: string
-  from: string
+  collapse_key: string;
+  from: string;
   notification:
-    {
-      body: string
-      title: string
-      icon: string
-    }
+  {
+    body: string
+    title: string
+    icon: string
+  };
 
 }
