@@ -1,38 +1,26 @@
-import {fuelTypes, User} from "./universal";
+import { fuelTypes, User } from "./universal";
+import { CustomerContact } from "./Customer";
+import { Truck, Batch, emptytruck } from "./Truck";
 
-export interface Order_ {
-  Id: string, //used to temporarily store the key, used later for looping
+export interface Order {
+  Id: string; // used to temporarily store the key, used later for looping
   company: {
     name: string,
     Id: string,
     phone: string,
-    contact: {
-      email: string,
-      name: string,
-      phone: string
-    },
+    contact: Array<CustomerContact>;
     krapin: string,
     QbId: string,
-    email?: string
   };
-  QbId: string,
-  InvoiceId: string,
-  stage: number,
-  origin: string,
-  /**
-   * @deprecated use QbID or Id instead
-   */
-  orderId: number,
+  QbId: string;
+  stage: number;
+  origin: string;
   notifications: {
-    allowsms: boolean,
-    allowemail: boolean
+    sms: boolean,
+    email: boolean
   };
   config: {
-    depot: {
-      name: string,
-      Id: string,
-    },
-    companyid: string,
+    depotId: string,
     sandbox: boolean,
   };
   error?: {
@@ -45,7 +33,7 @@ export interface Order_ {
   /**
    *@Deprecated, In Daudi 3 the truck Id is exactly the same as the order Id since its a 1:1 r/ship
    */
-  truck?: Array<string>;
+  truck: Truck;
   loaded: boolean;
   fuel: {
     [key in fuelTypes]: fuelconfig
@@ -68,11 +56,14 @@ export interface Order_ {
   };
 }
 
-export type fuelconfig = {
-  QbId: string,
-  qty: number,
-  priceconfig: priceconfig
-};
+
+
+export interface fuelconfig {
+  QbId: string;
+  qty: number;
+  priceconfig: priceconfig;
+  batches: Array<Batch>;
+}
 
 const initorderfuel = {
   qty: 0,
@@ -102,66 +93,62 @@ const initstages = {
   }
 };
 
-export type priceconfig = {
+export interface priceconfig {
   /**
    * @description the total price of the fuel/l, inclusive of VAT
    */
-  price: number,
+  price: number;
   /**
    * @description the amount/l without tax
    */
-  nonTaxprice: number,
+  nonTaxprice: number;
   /**
    * @description the amount/l that is not taxed, provided by the Gov
    */
-  nonTax: number,
+  nonTax: number;
   /**
    * @description the price set as the day's selling price used for discount calculation
    */
-  retailprice: number,
+  retailprice: number;
   /**
    * @description the minSp as of writing the order, connected to the buying of the most recent batch number
    */
-  minsp: number,
+  minsp: number;
   /**
    * @description the total amount of money that the order will cost in KES
    */
-  total: number,
+  total: number;
   /**
    * @description Amount of tax in the order in KES
    */
-  taxAmnt: number,
+  taxAmnt: number;
   /**
    * @description Worth of the order without tax
    */
-  nonTaxtotal: number,
+  nonTaxtotal: number;
   /**
    * @description calculation of price minus nonTax
    */
-  taxablePrice: number,
+  taxablePrice: number;
   /**
    * @description total amount on which tax is calculated on
    */
-  taxableAmnt: number,
+  taxableAmnt: number;
   /**
    * @description calculation of the price minus retailprice used in calculating the amount of discount
    * +ve for upmark and -ve for discount
    */
-  difference: number,
+  difference: number;
   /**
    * @description QuickbooksId of the tax Object, used for referencing for accounting
    */
-  taxQbId: string
-};
+  taxQbId: string;
+}
 
-export const emptyorder: Order_ = {
+export const emptyorder: Order = {
   Id: null,
   company: {
-    contact: {
-      email: null,
-      name: null,
-      phone: null
-    },
+    contact: [],
     phone: null,
     name: null,
     Id: null,
@@ -169,30 +156,25 @@ export const emptyorder: Order_ = {
     krapin: null
   },
   QbId: null,
-  InvoiceId: null,
+  truck: { ...emptytruck },
   notifications: {
-    allowsms: null,
-    allowemail: null
+    sms: null,
+    email: null
   },
   config: {
-    depot: {
-      name: null,
-      Id: null
-    },
-    companyid: null,
+    depotId: null,
     sandbox: null
   },
-  orderId: null,
   origin: null,
   stage: null,
   loaded: null,
   stagedata: {
-    1: initstages,
-    2: initstages,
-    3: initstages,
-    4: initstages,
-    5: initstages,
-    6: initstages
+    1: { ...initstages },
+    2: { ...initstages },
+    3: { ...initstages },
+    4: { ...initstages },
+    5: { ...initstages },
+    6: { ...initstages }
   },
   discount: {
     approved: null,
@@ -220,7 +202,8 @@ export const emptyorder: Order_ = {
         taxableAmnt: 0,
         taxQbId: null
       },
-      QbId: null
+      QbId: null,
+      batches: []
     },
     ago: {
       qty: 0,
@@ -238,7 +221,8 @@ export const emptyorder: Order_ = {
         taxableAmnt: 0,
         taxQbId: null
       },
-      QbId: null
+      QbId: null,
+      batches: []
     },
     ik: {
       qty: 0,
@@ -254,13 +238,14 @@ export const emptyorder: Order_ = {
         nonTaxtotal: 0,
         taxablePrice: 0,
         taxableAmnt: 0,
-        taxQbId: null
+        taxQbId: null,
       },
-      QbId: null
+      QbId: null,
+      batches: []
     }
   }
 };
-export type orderStages = "1" | "2" | "3" | "4" | "5" | "6"
+export type orderStages = "1" | "2" | "3" | "4" | "5" | "6";
 export let orderStagesarray = ["1", "2", "3", "4", "5", "6"];
 export let orderqueryStagesarray = ["1", "2", "3", "4", "6"];
 
